requirements.txt
fastapi
numpy
tensorflow
pillow
fastapi-socketio